refactor: CarouselSkeleton 컴포넌트 skeletons 폴더로 이동
