+++ src/pages/main/components/skeletons/SkeletonComponent.tsx
@@ -1,4 +1,4 @@
-import CarouselSkeleton from '../CarouselSkeleton';
+import CarouselSkeleton from './CarouselSkeleton';
 
 interface groupLengthPropTypes {
   groupLength: number;
